+++ lib/td_dq/rules/rules.ex
@@ -439,6 +439,17 @@ defmodule TdDq.Rules do
   end
 
   @doc """
+    Returns last rule_result for each rule_implementation of rule
+  """
+  def get_last_rule_implementations_result(rule) do
+    rule
+    |> Repo.preload(:rule_implementations)
+    |> Map.get(:rule_implementations)
+    |> Enum.map(&get_last_rule_result(&1.implementation_key))
+    |> Enum.filter(& &1)
+  end
+
+  @doc """
   Returns the list of rule_implementations.
 
   ## Examples
@@ -513,6 +524,18 @@ defmodule TdDq.Rules do
     |> join(:inner, [ri], r in assoc(ri, :rule))
     |> where([_, r], is_nil(r.deleted_at))
     |> Repo.get_by!(implementation_key: implementation_key)
+  end
+
+  def get_rule_by_implementation_key(implementation_key) do
+    implementation_rule =
+      implementation_key
+      |> get_rule_implementation_by_key()
+      |> Repo.preload(rule: :rule_type)
+
+    case implementation_rule do
+      nil -> nil
+      _rule -> Map.get(implementation_rule, :rule)
+    end
   end
 
   @doc """
@@ -915,6 +938,7 @@ defmodule TdDq.Rules do
   defp to_schema_type("string"), do: :string
   defp to_schema_type("list"), do: {:array, :string}
   defp to_schema_type("date"), do: :string
+  defp to_schema_type("structure"), do: :map
 
   def check_available_implementation_key(%{"implementation_key" => ""}),
     do: {:implementation_key_available}