+++ lib/td_dq_web/controllers/rule_result_controller.ex
@@ -8,6 +8,8 @@ defmodule TdDqWeb.RuleResultController do
   alias TdDq.Repo
   alias TdDq.Rules
 
+  @search_service Application.get_env(:td_dq, :elasticsearch)[:search_service]
+
   @rules_results_query ~S"""
     INSERT INTO rule_results ("implementation_key", "date", "result", parent_domains, inserted_at, updated_at)
     VALUES ($1, $2, $3, $4, $5, $5)
@@ -16,7 +18,7 @@ defmodule TdDqWeb.RuleResultController do
 
   # TODO: tets this
   def upload(conn, params) do
-    do_upload(conn, params)
+    do_upload(params)
     send_resp(conn, :ok, "")
   rescue
     e in RuntimeError ->
@@ -24,28 +26,52 @@ defmodule TdDqWeb.RuleResultController do
       send_resp(conn, :unprocessable_entity, Poison.encode!(%{error: e.message}))
   end
 
-  defp do_upload(conn, params) do
+  defp do_upload(params) do
     Logger.info("Uploading rule results...")
 
     start_time = DateTime.utc_now()
-    rules_results_upload = Map.get(params, @rule_results_param)
 
-    Repo.transaction(fn ->
-      upload_in_transaction(conn, rules_results_upload.path)
-    end)
+    rule_results_data =
+      params
+      |> Map.get(@rule_results_param)
+      |> rule_results_from_csv()
+
+    with {:ok, _} <- upload_data(rule_results_data) do
+      index_rule_results(rule_results_data)
+    end
 
     end_time = DateTime.utc_now()
 
     Logger.info("Metadata uploaded. Elapsed seconds: #{DateTime.diff(end_time, start_time)}")
   end
 
-  defp upload_in_transaction(_conn, rules_results_upload_path) do
-    Logger.info("Uploading rule results...")
-
-    rules_results_upload_path
+  defp rule_results_from_csv(%{path: path}) do
+    path
     |> File.stream!()
     |> Stream.drop(1)
     |> CSV.decode!(separator: ?;)
+  end
+
+  defp upload_data(rule_results_data) do
+    Repo.transaction(fn ->
+      upload_in_transaction(rule_results_data)
+    end)
+  end
+
+  defp index_rule_results(rule_results_data) do
+    rule_results_data
+    |> Enum.map(fn [implementation_key | _] ->
+      Rules.get_rule_by_implementation_key(implementation_key)
+    end)
+    |> Enum.filter(&(not is_nil(&1)))
+    |> Enum.uniq_by(fn %{id: id} -> id end)
+    |> Enum.map(&@search_service.put_searchable(&1))
+  end
+
+  defp upload_in_transaction(rules_results) do
+    Logger.info("Uploading rule results...")
+
+    rules_results
     |> Enum.each(fn data ->
       data =
         List.update_at(data, 1, fn x ->