+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## [Unreleased] 2019-06-18
+## [Unreleased]
+
+### Added
+
+- [TD-1533] Added structure schema type for system params
+- [TD-1690] index implementation results on elastic
+
+### Removed
+
+- [TD-1902] Removed description field from RuleImplementation
+
+### Changed
 
 - [TD-1893] Use CI_JOB_ID instead of CI_PIPELINE_ID